third_party/blink/web_tests/platform/generic/fast/css/readwrite-contenteditable-recalc-expected.txt

PASS getComputedStyle(p1)["background-color"] is "rgb(255, 204, 204)"
PASS getComputedStyle(c1a)["background-color"] is "rgb(255, 204, 204)"
PASS getComputedStyle(c1b)["background-color"] is "rgb(255, 204, 204)"
PASS getComputedStyle(p2)["background-color"] is "rgb(204, 255, 255)"
PASS getComputedStyle(c2a)["background-color"] is "rgb(204, 255, 255)"
PASS getComputedStyle(c2b)["background-color"] is "rgb(204, 255, 255)"
PASS getComputedStyle(p1)["background-color"] is "rgb(204, 255, 255)"
PASS getComputedStyle(c1a)["background-color"] is "rgb(204, 255, 255)"
PASS getComputedStyle(c1b)["background-color"] is "rgb(204, 255, 255)"
PASS getComputedStyle(p2)["background-color"] is "rgb(255, 204, 204)"
PASS getComputedStyle(c2a)["background-color"] is "rgb(255, 204, 204)"
PASS getComputedStyle(c2b)["background-color"] is "rgb(255, 204, 204)"
PASS successfullyParsed is true

TEST COMPLETE

Some parent text. Some child text.
Some parent text. Some child text.
